+++ chroma-common/tests/blockdevices/test_blockdevice_linux.py
@@ -1,7 +1,7 @@
 import mock
 
 from chroma_common.blockdevices.blockdevice_linux import BlockDeviceLinux
-from tests.common.blockdevices.blockdevice_base_tests import BaseTestBD
+from tests.blockdevices.blockdevice_base_tests import BaseTestBD
 from chroma_common.test.command_capture_testcase import CommandCaptureCommand
