+++ chroma-common/tests/blockdevices/test_blockdevice_zfs.py
@@ -3,8 +3,8 @@ import mock
 
 from os import path
 from chroma_common.blockdevices.blockdevice_zfs import BlockDeviceZfs
-from tests.common.blockdevices.blockdevice_base_tests import BaseTestBD
-from tests.data.chroma_common import example_data
+from tests.blockdevices.blockdevice_base_tests import BaseTestBD
+from tests.data import example_data
 from chroma_common.test.command_capture_testcase import CommandCaptureCommand
