+++ chroma-common/tests/filesystems/test_filesystem_ldiskfs.py
@@ -3,7 +3,7 @@ import mock
 from chroma_common.test.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
 from chroma_common.filesystems.filesystem_ldiskfs import FileSystemLdiskfs
 from chroma_common.blockdevices.blockdevice_linux import BlockDeviceLinux
-from tests.data.chroma_common import example_data
+from tests.data import example_data
 
 
 class TestFileSystemLdiskfs(CommandCaptureTestCase):
